Moved plain render routes to html-routes file
