+++ controllers/login-routes.js
@@ -1,18 +1,7 @@
 
 module.exports = function (app, passport) {
 
-  // HOME PAGE (with login links) ========
-  app.get('/', function (req, res) {
-    res.render('index.ejs'); // load the index.ejs file
-  });
-
   // LOGIN ===============================
-  // show the login form
-  app.get('/login', function (req, res) {
-    // render the page and pass in any flash data if it exists
-    res.render('login.ejs', { message: req.flash('loginMessage') });
-  });
-
   // process the login form
   app.post('/login', passport.authenticate('local-login', {
 		successRedirect : '/profile', // redirect to the secure profile section
@@ -20,13 +9,8 @@ module.exports = function (app, passport) {
 		failureFlash : true // allow flash messages
   }));
 
-  // SIGNUP ==============================
-  // show the signup form
-  app.get('/signup', function (req, res) {
-    // render the page and pass in any flash data if it exists
-    res.render('signup.ejs', { message: req.flash('signupMessage') });
-  });
 
+  // SIGNUP ==============================
   // process the signup form
   app.post('/signup', passport.authenticate('local-signup', {
 		successRedirect : '/profile', // redirect to the secure profile section
@@ -35,33 +19,9 @@ module.exports = function (app, passport) {
 	}));
 
 
-  // PROFILE SECTION =====================
-  // we will want this protected so you have to be logged in to visit
-  // we will use route middleware to verify this (the isLoggedIn function)
-  app.get('/profile', isLoggedIn, function (req, res) {
-    console.log('./controllers/login-routes.js - GET /profile ===============');
-    console.log('req.user: ', req.user);
-
-    res.render('profile.ejs', {
-      user: req.user[0] // get the user out of session and pass to template
-    });
-  });
-
   // LOGOUT ==============================
   app.get('/logout', function (req, res) {
     req.logout();
     res.redirect('/');
   });
 };
-
-// route middleware to make sure a user is logged in
-function isLoggedIn(req, res, next) {
-
-  // if user is authenticated in the session, carry on 
-  if (req.isAuthenticated()) {
-    return next();
-  }
-
-  // if they aren't redirect them to the home page
-  res.redirect('/');
-}
